+++ src/client/java/com/ricedotwho/rsm/ui/clickgui/impl/module/settings/impl/MultiBoolValueComponent.java
@@ -2,6 +2,7 @@ package com.ricedotwho.rsm.ui.clickgui.impl.module.settings.impl;
 
 import com.ricedotwho.rsm.data.StopWatch;
 import com.ricedotwho.rsm.ui.clickgui.api.FatalityColors;
+import com.ricedotwho.rsm.ui.clickgui.api.Mask;
 import com.ricedotwho.rsm.ui.clickgui.impl.module.ModuleComponent;
 import com.ricedotwho.rsm.ui.clickgui.impl.module.settings.ValueComponent;
 import com.ricedotwho.rsm.ui.clickgui.settings.impl.MultiBoolSetting;
@@ -27,7 +28,7 @@ public class MultiBoolValueComponent extends ValueComponent<MultiBoolSetting> {
     private final Map<String, Boolean> hoverStates = new HashMap<>();
 
     @Override
-    public void render(int mouseX, int mouseY, float partialTicks) {
+    public void render(double mouseX, double mouseY, float partialTicks) {
         float posX = getPosition().x;
         float posY = getPosition().y;
         float rectWidth = 100;
@@ -98,7 +99,7 @@ public class MultiBoolValueComponent extends ValueComponent<MultiBoolSetting> {
     }
 
     @Override
-    public void click(int mouseX, int mouseY, float mouseButton) {
+    public void click(double mouseX, double mouseY, float mouseButton) {
         float posX = getPosition().x;
         float posY = getPosition().y;
         float rectWidth = 100;
@@ -122,7 +123,7 @@ public class MultiBoolValueComponent extends ValueComponent<MultiBoolSetting> {
     }
 
     @Override
-    public void release(int mouseX, int mouseY, float mouseButton) {
+    public void release(double mouseX, double mouseY, float mouseButton) {
 
     }
 }
